Revert storing microseconds timestamp in .SST
This was written for better syncing with videos, but I have missed that
the mvhd creation time stores only seconds (even when type == 1, and
8-byte timestamps are used), so there's no point.

I have not removed tracking time with microseconds accuracy from the
fw code though. It does not affect anything, and it might come in handy
later.