Handle STATUS_FILE_SENT properly
This commit fixes two things regarding the STATUS_FILE_SENT status. It
is no longer set as soon as the entire file is sent, rather we wait for
the client the acknowledge that it has received the file. The second fix
is that we properly close the TCP connection after this ack from the
client arrives.