Prevent tcp_poll closing connection
According to the documentation, tcp_poll should fire when there was no
communication for a given period of time. In practice, sometimes I get
tcp_poll's while sending a file (actually sending, not just buffering,
since tcp_sent is fired constantly), which terminates a perfectly good
connection. This commit fixes this issue, but we run the risk of not
terminating actual bad connections.